phase2: migrate backend core to direct openai provider
