+++ backend/.env.example
@@ -1,28 +1,13 @@
 # ============================================
-# Azure OpenAI Realtime API Configuration
+# OpenAI Realtime API Configuration
 # ============================================
 # Required for voice-only interview feature
-# Get these from Azure Portal: Azure OpenAI resource
-# 1. Create Azure OpenAI resource in Azure Portal
-# 2. Deploy gpt-4o-realtime model
-# 3. Get API key and endpoint from resource overview
-AZURE_OPENAI_API_KEY=your-azure-openai-api-key-here
-AZURE_OPENAI_ENDPOINT=https://your-resource.openai.azure.com
-AZURE_OPENAI_DEPLOYMENT=gpt-4o-realtime
-# Separate deployment for chat completions used by feedback/report chains
-AZURE_OPENAI_CHAT_DEPLOYMENT=gpt-4o-mini
-AZURE_OPENAI_API_VERSION=2024-10-01-preview
-# Optional: realtime voice profile (for interview TTS). Example local value: ash
-REALTIME_VOICE=alloy
-
-# ============================================
-# Fallback: OpenAI Realtime API
-# ============================================
-# Use if Azure OpenAI is not available
-# Get from https://platform.openai.com/api-keys
-OPENAI_REALTIME_API_KEY=your-openai-api-key-here
-OPENAI_REALTIME_ENDPOINT=https://api.openai.com/v1/realtime
+OPENAI_API_KEY=your-openai-api-key-here
 OPENAI_REALTIME_MODEL=gpt-4o-realtime-preview
+OPENAI_CHAT_MODEL=gpt-4o-mini
+OPENAI_TRANSCRIBE_MODEL=gpt-4o-mini-transcribe
+# Optional: realtime voice profile
+REALTIME_VOICE=alloy
 
 # ============================================
 # Azure Speech Services
@@ -76,7 +61,7 @@ JWT_SECRET_KEY=
 # ============================================
 # Local example:
 # REDIS_URL=redis://localhost:6379
-# Azure Cache for Redis example (TLS):
+# Managed Redis example (TLS):
 # REDIS_URL=rediss://:PASSWORD@HOST:6380
 
 # ============================================
@@ -92,14 +77,5 @@ JWT_SECRET_KEY=
 # RATE_LIMIT_API_PER_MINUTE=300
 
 
-# ============================================
-# Azure AD Authentication
-# ============================================
-# For Azure AD-based authentication (alternative to API keys)
-AZURE_REALTIME_SCOPE=https://ai.azure.com/.default
-
-# ============================================
-# OpenAI API (Legacy)
-# ============================================
-# For other OpenAI features
-OPENAI_API_KEY=your-openai-api-key-here
+# Legacy compatibility aliases (optional)
+# OPENAI_REALTIME_API_KEY=${OPENAI_API_KEY}
