+++ backend/check-env.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""Check if Azure OpenAI keys are configured in .env file"""
+"""Check if OpenAI keys are configured in .env file."""
 import os
 import pathlib
 from dotenv import load_dotenv
@@ -14,30 +14,28 @@ else:
     exit(1)
 
 # Check keys
-api_key = os.getenv('AZURE_OPENAI_API_KEY', '')
-endpoint = os.getenv('AZURE_OPENAI_ENDPOINT', '')
+api_key = os.getenv('OPENAI_API_KEY', '') or os.getenv('OPENAI_REALTIME_API_KEY', '')
+model = os.getenv('OPENAI_REALTIME_MODEL', '')
 
 print("\n" + "="*60)
-print("AZURE OPENAI CONFIGURATION CHECK")
+print("OPENAI CONFIGURATION CHECK")
 print("="*60)
 
-if api_key and api_key != 'your-azure-openai-api-key-here' and len(api_key) > 10:
-    print(f"✅ AZURE_OPENAI_API_KEY: {api_key[:10]}...{api_key[-4:]} (length: {len(api_key)})")
+if api_key and api_key != 'your-openai-api-key-here' and len(api_key) > 10:
+    print(f"✅ OPENAI_API_KEY: {api_key[:10]}...{api_key[-4:]} (length: {len(api_key)})")
 else:
-    print(f"❌ AZURE_OPENAI_API_KEY: Not configured or using placeholder")
+    print(f"❌ OPENAI_API_KEY: Not configured or using placeholder")
     print("   Current value:", api_key[:50] if api_key else "EMPTY")
 
-if endpoint and endpoint != 'https://your-resource.openai.azure.com' and ('openai.azure.com' in endpoint or 'cognitiveservices.azure.com' in endpoint):
-    print(f"✅ AZURE_OPENAI_ENDPOINT: {endpoint}")
+if model:
+    print(f"✅ OPENAI_REALTIME_MODEL: {model}")
 else:
-    print(f"❌ AZURE_OPENAI_ENDPOINT: Not configured or using placeholder")
-    print("   Current value:", endpoint if endpoint else "EMPTY")
+    print(f"⚠️ OPENAI_REALTIME_MODEL: Not configured (will use server default)")
 
 print("\n" + "="*60)
-if api_key and api_key != 'your-azure-openai-api-key-here' and endpoint and endpoint != 'https://your-resource.openai.azure.com':
+if api_key and api_key != 'your-openai-api-key-here':
     print("✅ Configuration looks good!")
 else:
-    print("❌ Please update backend/.env with your actual Azure OpenAI keys:")
-    print("   AZURE_OPENAI_API_KEY=your-actual-key-here")
-    print("   AZURE_OPENAI_ENDPOINT=https://your-resource.openai.azure.com")
+    print("❌ Please update backend/.env with your actual OpenAI key:")
+    print("   OPENAI_API_KEY=your-actual-key-here")
 print("="*60)
