+++ backend/requirements.txt
@@ -1,4 +1,4 @@
-# Build tools (required for reliable Azure App Service deployment)
+# Build tools
 pip>=24.0
 setuptools>=70.0
 wheel>=0.41.0
@@ -30,9 +30,6 @@ email-validator>=2.1.0
 python-dotenv==1.0.0
 typing-extensions>=4.12.0
 python-dateutil==2.8.2
-
-# Azure services
-azure-identity==1.17.1
 
 # PDF generation
 reportlab==4.0.9
